[RENAME] common 디렉토리 변경
